% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lean.R
\name{add_lean}
\alias{add_lean}
\title{Plot lean in roi?}
\usage{
add_lean(building)
}
\arguments{
\item{building}{required, 8 digit building number}
}
\description{
This function computes ROI of building ECM retrofits, it needs time, cost, weather and energy
}
\examples{
roiBuilding(building="UT0032ZZ")
}
\keyword{roi}
