% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitting}
\alias{fitting}
\title{fitting models and plot fit to predict baseline consumption}
\usage{
fitting(building, dfPre, dfPost, colname, plotType, consumptionUnit, method,
  method_label, action_time, acc_result)
}
\arguments{
\item{building}{required, 8 digit building number}

\item{dfPre}{required, dataframe for pre retrofit energy}

\item{dfPost}{required, dataframe for post retrofit energy}

\item{colname}{column to use as the response variable}

\item{plotType}{elec or gas}

\item{consumptionUnit}{unit used for the response variable}

\item{method}{a function to fit the model of y ~ x}

\item{method_label}{string to label the method name}

\item{action_time}{retrofit time}

\item{acc_result}{accumulator data frame for fitting result}
}
\description{
This function predict the counterfactual electricity KWH or gas cubic foot consumption
}
\examples{
fitting(building=building, dfPre=dfPre, dfPost=dfPost, colname="Electricity_(KWH)", plotType="elec",
        consumptionUnit="KWH", method=lean.analysis::loess_fit, action_time=as.character(pre_end),
        method_label="loess", acc_result=acc_result)
}
\keyword{consumption}
\keyword{energy}
\keyword{fitting}
