% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeightedDaily.R
\name{getWeightedDaily}
\alias{getWeightedDaily}
\title{Get average temperature for a single building}
\usage{
getWeightedDaily(b, start_str, end_str, v, path = NULL)
}
\arguments{
\item{b}{required, 8 digit building number}

\item{start_str}{time to start downloading, a date string of yyyy-mm-dd format}

\item{end_str}{time to end downloading, a date string of yyyy-mm-dd format}

\item{v}{variable to compile}

\item{path}{optional, directory to save downloaded data.}
}
\description{
This function downloads 3 closest GHCND weather station data for variable v, and
computes a inverse weighted average of the variable
}
\examples{
getAvgTemp(building="UT0032ZZ", start_str="2012-01-05", end_str="2015-01-30")
}
\keyword{average}
\keyword{daily}
\keyword{noaa}
\keyword{temperature}
