% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{roiBuilding}
\alias{roiBuilding}
\title{compute ROI for building}
\usage{
roiBuilding(building)
}
\arguments{
\item{building}{required, 8 digit building number}
}
\description{
This function computes ROI of building ECM retrofits, it needs time, cost, weather and energy
}
\examples{
roiBuilding(building="UT0032ZZ")
}
\keyword{roi}
